moved interfaces to shared folder
